+++ server/Timelogger/Entities/TimeRegistrations.cs
@@ -4,7 +4,7 @@ namespace Timelogger.Entities
     {
         public int Id { get; set; }
         public int ProjectId { get; set; } 
-        public int RegisteredTime { get; set; }
+        public float RegisteredTime { get; set; }
         public string Note { get; set; }
         public string Date { get; set; }
     }
